Links should be target="_blank"
To make a link open in a new tab you need to set the target to _blank.
"blank" creates a new tab only for the first link clicked, and labels
that tab "blank", if you go back to the bento tab and click another link
with target="blank" it will navigate the previously opened tab to the
href of the link instead of opening in a new tab.