Adding functionality for dark and light theme to auto change
Adds functionality for dark and light theme to auto change based on time (6:00 pm and 5:00 am), and functionality to change based on system theme. Note, for the system theme, this was only tested on mac os. I'm not super well versed in Javascript, so I'm not entirely sure if it will work in other operating systems as well.